requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and model dependencies
transformers>=4.40.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1
audioread>=3.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
opencv-python>=4.8.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: For text paraphrasing
openai>=1.0.0  # If using GPT for paraphrasing
anthropic>=0.3.0  # Alternative for paraphrasing
